Fixed a bug that still produced correct behaviour
This was actually pretty interesting. I had the addresses wrong for
the DDR_B and PORT_B constants.

What it was doing was toggling the DDR_B bits, so switching between
input and output for the LED 'pin' is what created the blinking. I
wasn't sure if output or input was 'on' since I wasn't explicitly
turning it off or on. After just trying both, output was on.

I still haven't yet figured out how to dissassemble the files, so I
didn't know exactly which instructions rust emited, and thus what
address ranges to use to access some of the lower registers. But since
the addresses were wrong, I'm going to assume I want the higher
(+0x20) addresses from here, until I know better.